Revise the changes made in TestHelper:
ExpectNoDiagnosticsAsync and ExpectDiagnosticAsync assertions used to have messages with useful information. We want to keep it.
Check code in the helpers classes and make sure it is called by the test classes. If not (i.e. dead code), remove it.
Finally, replace the call to the obsolete methods with a call to the new methods.